requirements.txt
pytest==7.2.1
selene==2.0.0b16
allure-pytest==2.13.0
pydantic==1.10.5
python-dotenv==0.21.0
Appium-Python-Client==2.6.1